+++ odbc-api/src/buffers/text_row_set.rs
@@ -112,8 +112,7 @@ unsafe impl RowSetBuffer for TextRowSet {
 }
 
 unsafe impl ParameterCollection for &TextRowSet {
-
-    fn paramset_size(&self) -> u32 {
+    fn parameter_set_size(&self) -> u32 {
         *self.num_rows as u32
     }