+++ odbc-api/src/parameter_collection.rs
@@ -1,4 +1,4 @@
-use crate::{handles::Statement, Error, Parameter};
+use crate::{handles::Statement, Error, Input};
 
 mod tuple;
 
@@ -7,10 +7,9 @@ mod tuple;
 /// ODBC allows to place question marks (`?`) in the statement text as placeholders. For each such
 /// placeholder a parameter needs to be bound to the statement before executing it.
 pub unsafe trait ParameterCollection {
-
     /// Number of parameters in the collection. This can be different from the maximum batch size
-    /// a buffer may be able to hold. Returnig `0` will cause the the query not to be executed.
-    fn paramset_size(&self) -> u32;
+    /// a buffer may be able to hold. Returning `0` will cause the the query not to be executed.
+    fn parameter_set_size(&self) -> u32;
 
     /// # Safety
     ///
@@ -22,29 +21,28 @@ pub unsafe trait ParameterCollection {
 
 unsafe impl<T> ParameterCollection for T
 where
-    T: Parameter,
+    T: Input,
 {
-    fn paramset_size(&self) -> u32 {
+    fn parameter_set_size(&self) -> u32 {
         1
     }
 
     unsafe fn bind_parameters_to(&self, stmt: &mut Statement) -> Result<(), Error> {
-        self.bind_parameter_to(stmt, 1)?;
-        Ok(())
+        stmt.bind_input_parameter(1, self)
     }
 }
 
 unsafe impl<T> ParameterCollection for &[T]
 where
-    T: Parameter,
+    T: Input,
 {
-    fn paramset_size(&self) -> u32 {
+    fn parameter_set_size(&self) -> u32 {
         1
     }
 
     unsafe fn bind_parameters_to(&self, stmt: &mut Statement) -> Result<(), Error> {
         for (index, parameter) in self.iter().enumerate() {
-            parameter.bind_parameter_to(stmt, index as u16 + 1)?;
+            stmt.bind_input_parameter(index as u16 + 1, parameter)?;
         }
         Ok(())
     }